Caesar updates. Added one file and modified two
I uploaded a new file for the sonde data that reduced it to only the five parameters that will be shown on the dashboard card.

On the Upload Data page, I changed so that you now download the newly added file for Sonde Data. I also added the text "For more information about uploading data, visit our FAQ page."

Lastly, I added a paragraph and link on the FAQ page that explains what more metrics are available to be uploaded. I also added spaces before and after links by using &nbsp;